minor UI tweaks

+++ src/main/resources/app/containers/adminPage.js
@@ -88,9 +88,9 @@ class AdminPage extends React.PureComponent {
         <Typography variant="display3" gutterBottom>
           Items
         </Typography>
-        <SearchComponent value={this.state.itemSearchValue} onChange={this.searchItemOnChange.bind(this)}/>
+        <SearchComponent value={this.state.itemSearchValue} onChange={this.searchItemOnChange.bind(this)} />
         <button onClick={this.addItemClick.bind(this)}>add</button>
-        {this.state.showItemForm ? <CreateItemComponent submit={this.itemSubmitClick.bind(this)} categories={this.props.categories}/> : null}
+        {this.state.showItemForm ? <CreateItemComponent submit={this.itemSubmitClick.bind(this)} categories={this.props.categories} /> : null}
 
         <Paper>
           <Table>
@@ -103,20 +103,22 @@ class AdminPage extends React.PureComponent {
                 <TableCell>Category</TableCell>
                 <TableCell>Id</TableCell>
                 <TableCell>Visible</TableCell>
+                <TableCell>Delete</TableCell>
+                <TableCell>Edit</TableCell>
               </TableRow>
             </TableHead>
             <TableBody>
               {this.props.items.map(item => {
-                if(item.name.toUpperCase().includes(this.state.itemSearchValue.toUpperCase())||
-                  item.categoryName.toUpperCase().includes(this.state.categorySearchValue.toUpperCase())
-                ){
-                  return <AdminItemComponent
-                    item={item}
-                    key={item.id}
-                    remove={this.props.deleteItem}
-                    visible={item.visible}
-                    toggleVisible={this.props.toggleItemVisible}
-                  />
+                if (item.name
+                    .toUpperCase()
+                    .includes(
+                      this.state.itemSearchValue.toUpperCase()
+                    ) || item.category
+                    .toUpperCase()
+                    .includes(
+                      this.state.categorySearchValue.toUpperCase()
+                    )) {
+                  return <AdminItemComponent item={item} key={item.id} remove={this.props.deleteItem} visible={item.visible} toggleVisible={this.props.toggleItemVisible} />;
                 }
               })}
             </TableBody>
@@ -127,11 +129,11 @@ class AdminPage extends React.PureComponent {
           Categories
         </Typography>
 
-        <SearchComponent value={this.state.categorySearchValue} onChange={this.searchCategoryOnChange.bind(this)}/>
+        <SearchComponent value={this.state.categorySearchValue} onChange={this.searchCategoryOnChange.bind(this)} />
         <button onClick={this.addCategoryClick.bind(this)}>
           add Category
         </button>
-        {this.state.showCategoryForm ? <CreateCategoryComponent submit={this.categorySubmitClick.bind(this)}/> : null}
+        {this.state.showCategoryForm ? <CreateCategoryComponent submit={this.categorySubmitClick.bind(this)} /> : null}
 
         <Paper>
           <Table>
@@ -140,24 +142,23 @@ class AdminPage extends React.PureComponent {
                 <TableCell>Title</TableCell>
                 <TableCell>Filter</TableCell>
                 <TableCell>Visible</TableCell>
+                <TableCell>Delete</TableCell>
+                <TableCell>Edit</TableCell>
               </TableRow>
             </TableHead>
             <TableBody>
               {this.props.categories.map(category => {
-                if(category.title.toUpperCase().includes(this.state.categorySearchValue.toUpperCase())){
-                  return <CategoryComponent
-                      category={category}
-                      key={category.id}
-                      remove={this.props.deleteCategory}
-                      visible={category.visible}
-                      toggleVisible={this.props.toggleCategoryVisible}
-                    />
+                if (category.title
+                    .toUpperCase()
+                    .includes(
+                      this.state.categorySearchValue.toUpperCase()
+                    )) {
+                  return <CategoryComponent category={category} key={category.id} remove={this.props.deleteCategory} visible={category.visible} toggleVisible={this.props.toggleCategoryVisible} />;
                 }
               })}
             </TableBody>
           </Table>
         </Paper>
-
       </div>;
   }
 }
